[refactor] Extract single test file for `CycleIterator`
